chore: moved REQUIRED_KEYS and check_env_keys and load_dotenv out of the test file
